WIP: First full draft.
Probably need to add a little more text here and there. have not touched on R(A)
yet. But the first meat on bones version ready for people to at least give some
feedback on.

Also added a .gitignore and will test this version to make sure it can compile
on other systems.